Improve DNS-SD RFC conformance for properties
According to RFC 6763, the properties encoded in TXT records have ASCII
string keys, but the values are binary data. Additionally, the keys are
case-insensitive. A new class (CaseInsensitiveDict) was added for
storing properties. The only constraint is that the keys are strings, so
it's used for both the properties as decoded from the TXT record, and
and for the properties after the values have been stringified. To keep
it simpler, just the minimal methods have been overridden, and they're
simply lower-casing the keys before calling the base dict methods.

Using an actual case-insensitive dict from some other library might be a
better call later on.